small corrections

+++ redline/views.py
@@ -1,5 +1,6 @@
 from .models import Car
 from .serializers import CarSerializer
+from .serializers import CarPostSerializer
 from rest_framework.views import APIView
 from rest_framework.response import Response
 from rest_framework import status
@@ -16,6 +17,13 @@ class CarView(APIView):
         serializer = CarSerializer(cars, many=True)
         return Response(serializer.data)
 
+    def post(self, request, version, format=None):
+        serializer = CarPostSerializer(data=request.data)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data, status=status.HTTP_201_CREATED)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
 
 class CarObjectView(APIView):
     """
@@ -24,33 +32,28 @@ class CarObjectView(APIView):
     PUT - Updates a single cars information
     DELETE - Removes a car from the list
     """
-    def get_object(self, uuid):
+    def get_object(self, id):
         try:
-            return Car.objects.get(uuid=uuid)
+            return Car.objects.get(id=id)
         except Car.DoesNotExist:
-            raise Http404
+            return None
 
-    def get(self, request, version, uuid, format=None):
-        car = self.get_object(uuid)
+    def get(self, request, version, id, format=None):
+        car = self.get_object(id)
         serializer = CarSerializer(car)
         return Resonse(serializer.data)
 
-    def post(self, request, version, format=None):
-        serializer = CarSerializer(data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data, status=status.HTTP_201_CREATED)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-
-    def put(self, request, version, uuid, format=None):
-        car = self.get_object(uuid)
-        serializer = CarSerializer(pilot, data=request.data)
+    def put(self, request, version, id, format=None):
+        car = self.get_object(id)
+        serializer = CarSerializer(car, data=request.data)
         if serializer.is_valid():
             serializer.save()
             return Response(serializer.data, status=status.HTTP_200_OK)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
-    def delete(self, request, version, uuid, format=None):
-        car = self.get_object(uuid)
-        car.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+    def delete(self, request, version, id, format=None):
+        car = self.get_object(id)
+        if car:
+            car.delete()
+            return Response(car, status=status.HTTP_200_OK)
+        return Response(status.HTTP_404_NOT_FOUND)
